Review: Move rooting logic to TestSiteLayoutConfiguration
